replace render.cpp
